% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaTransformations.R
\name{ApplyThetaTransformations}
\alias{ApplyThetaTransformations}
\title{Title}
\usage{
ApplyThetaTransformations(transform_functions, inverse = FALSE)
}
\arguments{
\item{inverse}{}
}
\value{

}
\description{
Title
}
