% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as.stanfit}
\alias{as.stanfit}
\title{Convert a VARGrowthFit into a stanfit object}
\usage{
as.stanfit(obj)
}
\arguments{
\item{obj}{A VARGrowthFit Object}
}
\value{
the stan fit object
}
\description{
Convert a VARGrowthFit into a stanfit object
}
