% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsMeanTrendFunctions.R
\name{BoundedGompertzEstFunction_nlme}
\alias{BoundedGompertzEstFunction_nlme}
\title{Estimates Initial values for a Gompertz trend using a non-linear mixed effects model.
Initial values are estimated with the usual non-truncated normal likelihood and bounds are simply 10% above or below the minimum.}
\usage{
BoundedGompertzEstFunction_nlme(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Estimates Initial values for a Gompertz trend using a non-linear mixed effects model.
Initial values are estimated with the usual non-truncated normal likelihood and bounds are simply 10% above or below the minimum.
}
