% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionCriterion.R
\name{cdic}
\alias{cdic}
\title{Conditional Deviance Information Criterion calculation for VARGrowthFit Posterior}
\usage{
cdic(object)
}
\arguments{
\item{object}{A VARGrowthFit Object}
}
\value{
A VARGrowthCriteria list object with cdic as the criteria, deviance of the posterior mean, mean of the deviances, and effective number of parameters.
}
\description{
Calculate the conditional deviance information criterion (cdic) for a VARGrowthFit object. Lower values indicated a better fit.
}
