% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{checkSamplerDiagnostics}
\alias{checkSamplerDiagnostics}
\title{Crude check of HMC sampler}
\usage{
checkSamplerDiagnostics(object)
}
\arguments{
\item{object}{an object that can be coerced to a stanfit object}
}
\value{
TRUE if it passes tests, FALSE if not
}
\description{
This function will check the number of divergent transitions is 0, whether bfmi is greater than or equal to 0.3,
and whether any transitions met the maximum specified tree depth. If any one of these checks are FALSE, then FALSE is returned.
Otherwise, true is returned. Please refer to stan documentation for more information on these diagnostics and how to fix them.
}
