% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsMeanTrendFunctions.R
\name{CreateObsMeanObj}
\alias{CreateObsMeanObj}
\title{Constructor for an Obs Mean Trend}
\usage{
CreateObsMeanObj(
  num_params,
  est_function,
  stan_input,
  mean_func,
  prior_params,
  sim_params,
  model,
  param_names = "",
  ...
)
}
\arguments{
\item{num_params}{the number of parameters in your mean trend. Do not count any variance parameters.}

\item{est_function}{A function for estimating the parameters of your model. Must return a named list with "fit" for the fitted object and "params" that extract all parameters estimates (including variance) from the model.}

\item{mean_func}{A function factory that will take in all the parameters of theta and return the parametric function with argument x for time.}

\item{prior_params}{Optional vector of the name of parameters that require priors}

\item{sim_params}{Option vector of names of parameters needed to run a simulation}

\item{model}{A model name for the parameterization}

\item{param_names}{Names of the parameters theta (not the variance parameters)}

\item{...}{Other meta data to be attached to the object}

\item{generator}{a function to generate simulations from the likelihood}
}
\value{
An ObsMeanDef object which is a list of all the given arguments.
}
\description{
This is not ready for users yet.
}
