% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosteriorFunctions.R
\name{CreateParamList}
\alias{CreateParamList}
\title{Format list of parameters of each posterior sample}
\usage{
CreateParamList(object, posterior, ...)
}
\arguments{
\item{object}{A VARGrowthFit object}

\item{...}{}
}
\value{

}
\description{
Format list of parameters of each posterior sample
}
