% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaTrends.R
\name{CreateThetaTrendObj}
\alias{CreateThetaTrendObj}
\title{Title}
\usage{
CreateThetaTrendObj(
  num_params,
  est_input,
  est_function,
  sim_function,
  stan_input,
  calcTheta,
  prior_params,
  sim_params,
  model,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
