% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaTrends.R
\name{DeterministicLinearModelTrend}
\alias{DeterministicLinearModelTrend}
\title{Specify a Linear model with no variance}
\usage{
DeterministicLinearModelTrend(data, formula = ~1, XTheta = NULL)
}
\arguments{
\item{data}{A VARGrowth data object}

\item{formula}{a formula specifying how to construct the linear model matrix}

\item{XTheta}{An optional design matrix with number of rows = number of groups in data}
}
\value{

}
\description{
Use this function to define a linear model mean for the group trends of the form \eqn{g(\theta_d) = X_\theta \beta_d}.
\eqn{X_\theta} should be a U x p matrix where U is the number of groups and p is the number of linear predictor parameters.
Note that as of now every mean parameter \eqn{\theta} is assumed to have the same design matrix.
}
