% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaTrends.R
\name{DeterministicLinearTrendEstFunction}
\alias{DeterministicLinearTrendEstFunction}
\title{Estimate Deterministic Linear trend of thetas}
\usage{
DeterministicLinearTrendEstFunction(theta, XTheta, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Estimate Deterministic Linear trend of thetas
}
