% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsMeanTrendFunctions.R
\name{GompertzEstFunction_nlme}
\alias{GompertzEstFunction_nlme}
\title{Estimates Initial values for a Gompertz trend using a non-linear mixed effects model}
\usage{
GompertzEstFunction_nlme(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Estimates Initial values for a Gompertz trend using a non-linear mixed effects model
}
