% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaTrends.R
\name{LinearTrendEstFunction2}
\alias{LinearTrendEstFunction2}
\title{Estimate Linear trend of thetas}
\usage{
LinearTrendEstFunction2(theta, XTheta, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Estimate Linear trend of thetas
}
