% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationFunctions.R
\name{linearTrendSim}
\alias{linearTrendSim}
\title{Title}
\usage{
linearTrendSim(
  U,
  XTheta,
  betaTheta,
  SigmaTheta,
  simplify = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
