% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationFunctions.R
\name{linearTrendSim2}
\alias{linearTrendSim2}
\title{Title}
\usage{
linearTrendSim2(
  U,
  XTheta,
  betaTheta,
  SigmaTheta,
  simplify = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
