% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionCriterion.R
\name{mdic}
\alias{mdic}
\title{Marginal Deviance Information Criterion calculation for VARGrowthFit Posterior}
\usage{
mdic(object, nsamp = 1e+05)
}
\arguments{
\item{object}{A VARGrowthFit Object}
}
\value{
A VARGrowthCriteria list object with cdic as the criteria, deviance of the posterior mean, mean of the deviances, and effective number of parameters.
}
\description{
Calculate the marginal deviance information criterion (mdic) for a VARGrowthFit object. The marginalization takes place over the theta values of the hierarchical model.Lower values indicate a better fit.
}
