% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosteriorFunctions.R
\name{plot.VARGrowthPosteriorPredictive}
\alias{plot.VARGrowthPosteriorPredictive}
\title{Plot Posterior predictive distribution}
\usage{
\method{plot}{VARGrowthPosteriorPredictive}(x, y = NULL, alpha = 0.05, estimate = mean, ...)
}
\arguments{
\item{x}{a VARGrowthPosteriorPredictive object}

\item{y}{argument is not used}

\item{alpha}{level to produce (1 - alpha) credible intervals}

\item{estimate}{a function to produce the point estimate of the posterior predictive distribution}

\item{...}{additional arguments that are not used (yet)}
}
\value{
a ggplot object that plots the point estimate and credible interval for each group
}
\description{
Plot Posterior predictive distribution
}
