% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosteriorFunctions.R, R/VARGrowth-package.R
\name{posterior_predict.VARGrowthFit}
\alias{posterior_predict.VARGrowthFit}
\alias{posterior_predict}
\title{Get Posterior Predictive distribution of VARGrowthFit}
\usage{
\method{posterior_predict}{VARGrowthFit}(
  object,
  newdata = NULL,
  ndraws = NULL,
  ncores = NULL,
  seed = sample.int(.Machine$integer.max, 1),
  ...
)
}
\arguments{
\item{object}{A VARGrowthFit object}

\item{newdata}{A VARGrowthData object}

\item{ndraws}{Number of draws to resample from the posterior distribution}

\item{ncores}{Number of cores to run in parallel. Not used if newdata = NULL}

\item{seed}{set seed for reproducibility}

\item{...}{additional arguments to be passed to posterior::resample_draws}
}
\value{
A matrix of size nrow(newdata) x ndraws
}
\description{
Get Posterior Predictive distribution of VARGrowthFit
}
