% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosteriorFunctions.R
\name{posterior_predict_UnobsGroups}
\alias{posterior_predict_UnobsGroups}
\title{Title}
\usage{
posterior_predict_UnobsGroups(
  object,
  param_list,
  newdata,
  ncores = NULL,
  simplify = TRUE
)
}
\arguments{
\item{simplify}{}
}
\value{

}
\description{
Title
}
