% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationFunctions.R
\name{simObs}
\alias{simObs}
\title{Title}
\usage{
simObs(
  dat,
  Theta,
  ObsVar,
  GrowthFunction,
  ThetaTransform,
  simplify = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
