% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationFunctions.R
\name{SimulateData}
\alias{SimulateData}
\title{Title}
\usage{
SimulateData(
  dat,
  param_list,
  ThetaTrend,
  GrowthFunction,
  ThetaTransform = NULL,
  seed = sample.int(.Machine$integer.max, 1),
  simplify = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
