% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosteriorFunctions.R
\name{summary.VARGrowthPosterior}
\alias{summary.VARGrowthPosterior}
\title{Title}
\usage{
\method{summary}{VARGrowthPosterior}(object, estimate = mean, alpha = 0.05, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
