% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARGrowthFit.R
\name{VARGrowth}
\alias{VARGrowth}
\title{Title}
\usage{
VARGrowth(
  data,
  ThetaTrend,
  GrowthFunction,
  PriorList,
  ThetaTransform = NULL,
  nchains = 4,
  seed = sample.int(.Machine$integer.max, 1),
  perturb_var = 10,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{verbose}{Whether to print information about the model being fit}

\item{...}{}
}
\value{

}
\description{
Title
}
