% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARGrowthFit.R
\name{VARGrowthInput}
\alias{VARGrowthInput}
\title{Title}
\usage{
VARGrowthInput(data, ThetaTrend, GrowthFunction, ThetaTransform, PriorList)
}
\arguments{
\item{PriorList}{}
}
\value{

}
\description{
Title
}
